backend/pyproject.toml
[project]
name = "ai-assessment-mvp"
version = "0.1.0"
description = "AI-powered assessment system with LangGraph agentic workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "python-multipart==0.0.9",
    "pydantic==2.8.2",
    "pydantic-settings==2.4.0",
    "langchain==0.3.0",
    "langchain-openai==0.2.0",
    "langchain-community==0.3.0",
    "langgraph==0.2.28",
    "openai==1.51.0",
    "supabase==2.7.4",
    "PyPDF2==3.0.1",
    "python-docx==1.1.2",
    "httpx==0.27.2",
    "typing>=3.10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
